change scatterplot material, use colour for different labels
Also removed .DS_Store from version control (these are files that
automatically pop up in every directory when using finder on a mac).

Main modifications are in both the GetData.cs file and the
ScatterPlot.cs file. Should fall back to the standard red points for
everything if there's no label column in the CSV